WoodCT-MultiModal/requirements.txt
numpy>=1.21.0
scipy>=1.7.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
torch>=1.12.0
torchvision>=0.13.0
Pillow>=9.0.0
tqdm>=4.62.0
